refactor: e2e 오류 해결
