+++ src/js/Operator.mjs
