+++ src/js/model/inputStore.mjs
@@ -1,0 +1,16 @@
+import { Operator } from "../Operator.mjs";
+
+export class InputStore extends Array {
+  constructor(...array) {
+    super(...array);
+    this.operator = new Operator();
+  }
+
+  push(item) {
+    if (this.operator.isOperator(item) || /^[0-9]$/.test(item)) {
+      super.push(item);
+      return this;
+    }
+    throw new Error(`not allowed item ${item}`);
+  }
+}
